+++ app/events/page.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import Card from "@/components/Card";
-import { events } from "@/public/data/events";
+import { events as oldEvents } from "@/public/data/events";
 import { Metadata } from "next";
+import Database from "@/appwrite/database";
 
 export const dynamic = "force-dynamic";
 
@@ -11,16 +12,22 @@ export const metadata: Metadata = {
   icons: { icon: "/images/abes-acm.png" },
 };
 
-const Events = () => {
+const Events = async () => {
+  const database = new Database();
+
+  const { documents } = await database.getEvents();
+
+  const events = [...documents, ...oldEvents];
+
   return (
     <main className="p-5 md:px-20 lg:px-36 space-y-5">
       <div className="flex space-x-1 items-end">
         <h1 className="text-4xl">Events</h1>{" "}
         <span className="opacity-75">hosted by us</span>
       </div>
-      <section className="gap-3 grid grid-cols-1 md:grid-cols-2 2xl:grid-cols-3">
+      <section className="gap-3 gap-y-5 md:gap-y-3 grid grid-cols-1 sm:grid-cols-2 2xl:grid-cols-3">
         {events.map((event, index) => (
-          <Card className={"w-10"} key={index} {...event} />
+          <Card className={"w-10"} key={index} {...event} image={event.logo} />
         ))}
       </section>
     </main>