+++ components/Card.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import Link from "next/link";
 import { ArrowRight } from "lucide-react";
+import Image from "next/image";
 
 interface CardProps {
   name: string;
@@ -15,27 +16,26 @@ const Card = ({ name, description, image, date, id, className }: CardProps) => {
   return (
     <Link
       href={`/events/${id}`}
-      className={`w-full md:w-80 lg:w-96 h-80 flex flex-col space-y-2 items-center bg-secondary/40 border rounded-md p-4 ${className}`}
+      className={`w-full sm:max-w-80 lg:w-full lg:max-w-full border rounded-xl overflow-clip flex flex-col lg:flex-row ${className}`}
     >
-      <div className="w-full justify-between flex flex-col space-y-1 items-start">
-        <h3 className="text-2xl flex items-center group space-x-1">
-          <span>{name}</span>
-          <span>
-            <ArrowRight
-              size={"1em"}
-              className="group-hover:translate-x-2 duration-150"
-            />
-          </span>
-        </h3>{" "}
-        <div className="flex items-start text-sm text-primary/90">{date}</div>
+      <Image
+        src={image!}
+        alt={name}
+        width={256}
+        height={256}
+        className={"w-full lg:w-1/2"}
+      />
+      <div className={"w-full h-full flex flex-col p-2 px-4 space-y-2"}>
+        <h3
+          className={
+            "text-2xl font-medium after:content-[''] flex after:bg-red-600"
+          }
+        >
+          {name}
+        </h3>
+        <span className={"opacity-75"}>{date}</span>
+        <p className={"opacity-90 "}>{description.slice(0, 75)}...</p>
       </div>
-      <br />
-      <hr className="h-0.5 w-full bg-red-50/30" />
-      <br />
-      {/* <hr className="h-[1.25px] bg-primary/25 w-full" /> */}
-      <p className="w-full text-sm overflow-hidden">
-        {description.split(" ").slice(0, 40).join(" ") + "..."}
-      </p>
     </Link>
   );
 };