+++ changelog.md
@@ -18,3 +18,9 @@ _Please refer to Git Commit History for previous Logs_
 ### 10 May 2024
 - Updated Date picker for Create Event Page
 - Added new default open graph image
+- Tweaks
+  - Modified Navbar
+  - Modified Homepage
+- Added delete event feature to admin panel
+- Fixed not being able to access admin panel correctly
+- Added Dialogs